Get rooms working
Made a lot of changes, but finally got the rooms working (I think). The roomnumber is now generated client side and send to the server with the emit. I have an empty erray which get filled with the newly made roomnumbers. I added validation and check if the room a user wants to join, is in the array.